+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Release Notes
 
+### New features
+* added output of dependent repositories and packages
+
 ## [Version 0.4.0](https://pypi.org/project/ghrepo-stats/0.4.0/) (2022-12-26)
 
 ### Changes